requirements.txt
aiolimiter==1.0.0
anyio==3.7.0
APScheduler==3.10.1
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
exceptiongroup==1.1.1
fake-useragent==1.1.3
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
outcome==1.2.0
packaging==23.1
PySocks==1.7.1
python-dotenv==1.0.0
python-telegram-bot==20.3
pytz==2023.3
requests==2.31.0
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
tornado==6.3.2
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.3
tzlocal==5.0.1
urllib3==2.0.3
webdriver-manager==3.8.6
wsproto==1.2.0
